Complete SIH Voice Activator demo
End-to-end working pipeline: real microphone -> KeywordDetector
(TinyKWSNet) -> AudioSessionManager -> ASR (faster-whisper) ->
CommandInterpreter -> device-state -> live WebSocket-driven UI.

- api/: FastAPI app, live-mode WebSocket session, device-state layer
- voice_pipeline/: orchestrator wiring wake -> capture -> ASR -> command
- keyword_spotting/: TinyKWSNet model, training pipeline, trained
  NOVA checkpoint (models/kws_nova_cnn.pt, ~31KB)
- asr/, audio_processing/, command/: transcription, mic capture/session
  management, rule-based command interpretation
- frontend/: photorealistic smart-room UI with per-device animated
  layers (light/fan/AC/TV/door) reacting to real device state
- tests/, scripts/, docs/, presentation/: full test suite, demo/debug
  scripts, stage documentation, SIH presentation deck

+++ tests/unit/test_conv1d_learn.py
@@ -10,8 +10,8 @@ D = 2
 print('Cin, Cout, F, S, B, Lin, D=', Cin, Cout, F, S, B, Lin, D)
 conv1d = torch.nn.Conv1d(Cin, Cin, F, S, bias=False, padding=(F-1)*D, dilation=D, groups=Cin)
 #conv1d = torch.nn.Conv1d(Cin, Cin, F, S, bias=False, padding=1, dilation=1, groups=Cin)
-inputs = torch.randint(3, (B, Cin, Lin))
-conv1d.weight.data = torch.randint(5, conv1d.weight.size())
+inputs = torch.randint(3, (B, Cin, Lin)).float()
+conv1d.weight.data = torch.randint(5, conv1d.weight.size()).float()
 outputs = conv1d(inputs)
 # Lout = (Lin - F) / S + 1
